@@ -1,4 +1,5 @@
-const {
+/* eslint-disable @typescript-eslint/ban-ts-comment */
+import {
   heroicConfigPath,
   heroicGamesConfigPath,
   launchGame,
@@ -19,27 +20,38 @@ const {
   showAboutWindow,
   kofiURL,
   heroicGithubURL,
-} = require('./utils')
+} from './utils'
 
-const byteSize = require('byte-size')
-const { spawn, exec } = require('child_process')
-const path = require('path')
-const isDev = require('electron-is-dev')
-const fs = require('fs')
-const promisify = require('util').promisify
+import byteSize from 'byte-size'
+import { spawn, exec } from 'child_process'
+import * as path from 'path'
+import isDev from 'electron-is-dev'
+import {
+  stat,
+  readFileSync,
+  readdirSync,
+  writeFile,
+  existsSync,
+  mkdirSync,
+} from 'fs'
+import { promisify } from 'util'
+import axios from 'axios'
+
 const execAsync = promisify(exec)
-const stat = promisify(fs.stat)
-const axios = require('axios')
+const statAsync = promisify(stat)
 
-const {
+import {
   app,
   BrowserWindow,
   ipcMain,
   Notification,
   Menu,
   Tray,
-  dialog: { showMessageBox },
-} = require('electron')
+  dialog,
+} from 'electron'
+import { Game, InstalledInfo, KeyImage } from './types.js'
+
+const showMessageBox = dialog.showMessageBox
 
 function createWindow() {
   // Create the browser window.
@@ -64,46 +76,43 @@ function createWindow() {
 
   //load the index.html from a url
   if (isDev) {
-    const {
-      default: installExtension,
-      REACT_DEVELOPER_TOOLS,
-    } = require('electron-devtools-installer')
+    import('electron-devtools-installer').then((devtools) => {
+      const { default: installExtension, REACT_DEVELOPER_TOOLS } = devtools
 
-    installExtension(REACT_DEVELOPER_TOOLS).catch((err) => {
-      console.log('An error occurred: ', err)
+      installExtension(REACT_DEVELOPER_TOOLS).catch((err: any) => {
+        console.log('An error occurred: ', err)
+      })
     })
-
     win.loadURL('http://localhost:3000')
     // Open the DevTools.
     win.webContents.openDevTools()
   } else {
     win.on('close', async (e) => {
       e.preventDefault()
-      //Send a warning if user want to close the window while some gaming is installing on background
-      const { response } = await showMessageBox({
-        title: 'Games Downloading',
-        message: 'Do you really want to quit? Downloads will be canceled',
-        buttons: ['NO', 'YES'],
-      })
-
-      if (response === 1) {
-        win.destroy()
-      }
+      win.hide()
     })
     win.loadURL(`file://${path.join(__dirname, '../build/index.html')}`)
     win.setMenu(null)
   }
 }
 
-// TODO: Update Legendary to latest version
-
 // This method will be called when Electron has finished
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
-let appIcon = null
+let appIcon: Tray = null
 app.whenReady().then(() => {
+  createWindow()
+
   appIcon = new Tray(icon)
+  const currentWindow: BrowserWindow = BrowserWindow.getAllWindows()[0]
+
   const contextMenu = Menu.buildFromTemplate([
+    {
+      label: 'Show Heroic',
+      click: function () {
+        currentWindow.show()
+      },
+    },
     {
       label: 'About',
       click: function () {
@@ -130,14 +139,10 @@ app.whenReady().then(() => {
     },
   ])
 
-  appIcon.setToolTip('Heroic')
   appIcon.setContextMenu(contextMenu)
-
-  return createWindow()
+  appIcon.setToolTip('Heroic')
+  return
 })
-
-app.on('ready', () => {})
-// Define basic paths
 
 ipcMain.on('Notify', (event, args) => {
   const notify = new Notification({
@@ -152,13 +157,13 @@ ipcMain.handle('writeFile', (event, args) => {
   const app = args[0]
   const config = args[1]
   if (args[0] === 'default') {
-    return fs.writeFile(
+    return writeFile(
       heroicConfigPath,
       JSON.stringify(config, null, 2),
       () => 'done'
     )
   }
-  return fs.writeFile(
+  return writeFile(
     `${heroicGamesConfigPath}/${app}.json`,
     JSON.stringify(config, null, 2),
     () => 'done'
@@ -179,6 +184,8 @@ ipcMain.handle('getGameInfo', async (event, game) => {
 })
 
 ipcMain.handle('launch', (event, appName) => {
+  console.log('launching', appName)
+
   return launchGame(appName).catch(console.log)
 })
 
@@ -220,7 +227,8 @@ ipcMain.handle('install', async (event, args) => {
   let command = `${legendaryBin} install ${game} --base-path '${path}' -y &> ${logPath}`
   if (path === 'default') {
     const { defaultInstallPath } = JSON.parse(
-      fs.readFileSync(heroicConfigPath)
+      // @ts-ignore
+      readFileSync(heroicConfigPath)
     ).defaultSettings
     command = `${legendaryBin} install ${game} --base-path ${defaultInstallPath} -y |& tee ${logPath}`
   }
@@ -244,7 +252,7 @@ ipcMain.on('requestGameProgress', (event, appName) => {
   const logPath = `${heroicGamesConfigPath}${appName}.log`
   exec(
     `tail ${logPath} | grep 'Progress: ' | awk '{print $5 $6}'`,
-    (error, stdout, stderr) => {
+    (error, stdout) => {
       const status = `${stdout.split('\n')[0]}`.split('(')
       const percent = status[0]
       const bytes = status[1] && status[1].replace('),', 'MB')
@@ -264,17 +272,14 @@ ipcMain.on('kill', (event, game) => {
 
 ipcMain.on('openFolder', (event, folder) => spawn('xdg-open', [folder]))
 
-ipcMain.on('getAlternativeWine', (event, args) =>
+ipcMain.on('getAlternativeWine', (event) =>
   event.reply('alternativeWine', getAlternativeWine())
 )
 
 // Calls WineCFG or Winetricks. If is WineCFG, use the same binary as wine to launch it to dont update the prefix
 ipcMain.on('callTool', async (event, { tool, wine, prefix }) => {
-  let wineBin = wine.replace('proton', 'dist/bin/wine64')
-
-  if (wine.endsWith('proton')) {
-    wineBin = wine.replace('proton', 'dist/bin/wine64')
-  }
+  const wineBin = wine.replace("/proton'", "/dist/bin/wine64'")
+  console.log({ wine, wineBin }, wine.endsWith("/proton'"))
 
   const command = `WINE=${wineBin} WINEPREFIX=${prefix} ${
     tool === 'winecfg' ? `${wineBin} ${tool}` : tool
@@ -289,13 +294,15 @@ ipcMain.on('requestSettings', (event, appName) => {
   if (appName !== 'default') {
     writeGameconfig(appName)
   }
-  const defaultSettings = JSON.parse(fs.readFileSync(heroicConfigPath))
+  // @ts-ignore
+  const defaultSettings = JSON.parse(readFileSync(heroicConfigPath))
   if (appName === 'default') {
     return event.reply('defaultSettings', defaultSettings.defaultSettings)
   }
-  if (fs.existsSync(`${heroicGamesConfigPath}${appName}.json`)) {
+  if (existsSync(`${heroicGamesConfigPath}${appName}.json`)) {
     settings = JSON.parse(
-      fs.readFileSync(`${heroicGamesConfigPath}${appName}.json`)
+      // @ts-ignore
+      readFileSync(`${heroicGamesConfigPath}${appName}.json`)
     )
     return event.reply(appName, settings[appName])
   }
@@ -320,12 +327,14 @@ ipcMain.handle('readFile', async (event, file) => {
   }
 
   const installed = `${legendaryConfigPath}/installed.json`
-  const files = {
-    user: loggedIn ? require(userInfo) : { displayName: null },
+  const files: any = {
+    // @ts-ignore
+    user: loggedIn ? JSON.parse(readFileSync(userInfo)) : { displayName: null },
     library: `${legendaryConfigPath}/metadata/`,
     config: heroicConfigPath,
-    installed: await stat(installed)
-      .then(() => JSON.parse(fs.readFileSync(installed)))
+    installed: await statAsync(installed)
+      // @ts-ignore
+      .then(() => JSON.parse(readFileSync(installed)))
       .catch(() => []),
   }
 
@@ -337,81 +346,83 @@ ipcMain.handle('readFile', async (event, file) => {
   }
 
   if (file === 'library') {
-    const library = fs.existsSync(files.library)
+    const library = existsSync(files.library)
     const fallBackImage =
       'https://user-images.githubusercontent.com/26871415/103480183-1fb00680-4dd3-11eb-9171-d8c4cc601fba.jpg'
 
     if (library) {
-      return fs
-        .readdirSync(files.library)
-        .map((file) => `${files.library}/${file}`)
-        .map((file) => JSON.parse(fs.readFileSync(file)))
-        .map(({ app_name, metadata }) => {
-          const {
-            description,
-            keyImages,
-            title,
-            developer,
-            customAttributes: { CloudSaveFolder },
-          } = metadata
-          const cloudSaveEnabled = Boolean(CloudSaveFolder)
-          const saveFolder = cloudSaveEnabled ? CloudSaveFolder.value : ''
-          const gameBox = keyImages.filter(
-            ({ type }) => type === 'DieselGameBox'
-          )[0]
-          const gameBoxTall = keyImages.filter(
-            ({ type }) => type === 'DieselGameBoxTall'
-          )[0]
-          const logo = keyImages.filter(
-            ({ type }) => type === 'DieselGameBoxLogo'
-          )[0]
+      return (
+        readdirSync(files.library)
+          .map((file) => `${files.library}/${file}`)
+          // @ts-ignore
+          .map((file) => JSON.parse(readFileSync(file)))
+          .map(({ app_name, metadata }) => {
+            const {
+              description,
+              keyImages,
+              title,
+              developer,
+              customAttributes: { CloudSaveFolder },
+            } = metadata
+            const cloudSaveEnabled = Boolean(CloudSaveFolder)
+            const saveFolder = cloudSaveEnabled ? CloudSaveFolder.value : ''
+            const gameBox = keyImages.filter(
+              ({ type }: KeyImage) => type === 'DieselGameBox'
+            )[0]
+            const gameBoxTall = keyImages.filter(
+              ({ type }: KeyImage) => type === 'DieselGameBoxTall'
+            )[0]
+            const logo = keyImages.filter(
+              ({ type }: KeyImage) => type === 'DieselGameBoxLogo'
+            )[0]
 
-          const art_cover = gameBox ? gameBox.url : null
-          const art_logo = logo ? logo.url : null
-          const art_square = gameBoxTall ? gameBoxTall.url : fallBackImage
+            const art_cover = gameBox ? gameBox.url : null
+            const art_logo = logo ? logo.url : null
+            const art_square = gameBoxTall ? gameBoxTall.url : fallBackImage
 
-          const installedGames = Object.values(files.installed)
-          const isInstalled = Boolean(
-            installedGames.filter((game) => game.app_name === app_name).length
-          )
-          const info = isInstalled
-            ? installedGames.filter((game) => game.app_name === app_name)[0]
-            : {}
+            const installedGames: Game[] = Object.values(files.installed)
+            const isInstalled = Boolean(
+              installedGames.filter((game) => game.app_name === app_name).length
+            )
+            const info = isInstalled
+              ? installedGames.filter((game) => game.app_name === app_name)[0]
+              : {}
 
-          const {
-            executable = null,
-            version = null,
-            install_size = null,
-            install_path = null,
-          } = info
+            const {
+              executable = null,
+              version = null,
+              install_size = null,
+              install_path = null,
+            } = info as InstalledInfo
 
-          const convertedSize = `${byteSize(install_size).value}${
-            byteSize(install_size).unit
-          }`
+            const convertedSize = `${byteSize(install_size).value}${
+              byteSize(install_size).unit
+            }`
 
-          return {
-            isInstalled,
-            info,
-            title,
-            executable,
-            version,
-            install_size: convertedSize,
-            install_path,
-            app_name,
-            developer,
-            description,
-            cloudSaveEnabled,
-            saveFolder,
-            art_cover: art_cover || art_square,
-            art_square: art_square || art_cover,
-            art_logo,
-          }
-        })
-        .sort((a, b) => {
-          const gameA = a.title.toUpperCase()
-          const gameB = b.title.toUpperCase()
-          return gameA < gameB ? -1 : 1
-        })
+            return {
+              isInstalled,
+              info,
+              title,
+              executable,
+              version,
+              install_size: convertedSize,
+              install_path,
+              app_name,
+              developer,
+              description,
+              cloudSaveEnabled,
+              saveFolder,
+              art_cover: art_cover || art_square,
+              art_square: art_square || art_cover,
+              art_logo,
+            }
+          })
+          .sort((a, b) => {
+            const gameA = a.title.toUpperCase()
+            const gameB = b.title.toUpperCase()
+            return gameA < gameB ? -1 : 1
+          })
+      )
     }
     return []
   }
@@ -431,15 +442,14 @@ ipcMain.handle('egsSync', async (event, args) => {
   return `${stdout} - ${stderr}`
 })
 
-// TODO: Check the best way to Sync saves to implement soon
 ipcMain.handle('syncSaves', async (event, args) => {
   const [arg = '', path, appName] = args
   const command = `${legendaryBin} sync-saves --save-path ${path} ${arg} ${appName} -y`
   const legendarySavesPath = `${home}/legendary/.saves`
 
   //workaround error when no .saves folder exists
-  if (!fs.existsSync(legendarySavesPath)) {
-    fs.mkdirSync(legendarySavesPath, { recursive: true })
+  if (!existsSync(legendarySavesPath)) {
+    mkdirSync(legendarySavesPath, { recursive: true })
   }
 
   console.log('\n syncing saves for ', appName)
